requirements.txt
# Core dependencies
confluent-kafka>=2.0.0
minio>=7.1.0
psycopg2-binary>=2.9.0
pandas>=1.3.0
numpy>=1.20.0
faker>=8.0.0

# Visualization
diagrams>=0.23.0
graphviz>=0.20.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-runner>=6.0.0
black>=23.0.0
flake8>=6.0.0